[refactor] 검색 기능 구현을 위해 Issue state을 위로 끌어올림
issue State을 Issue Page 컴포넌트로 끌어올려서 관리
